feat(core):支持kjs的startup脚本、优化模糊性能、一部分属性支持浮点数了

+++ src/main/java/com/sighs/apricityui/init/Style.java
@@ -571,7 +571,7 @@ public class Style implements Cloneable {
         );
     }
 
-    public record TextStroke(int width, int color) {
+    public record TextStroke(double width, int color) {
         public static final TextStroke NONE = new TextStroke(0, 0);
     }
